BACKEND DONE!

+++ KeeperFinal/Controllers/VaultKeepsController.cs
@@ -1,16 +1,18 @@
 namespace KeeperFinal.Controllers;
-
+[Authorize]
 [ApiController]
 [Route("api/[controller]")]
 public class VaultKeepsController : ControllerBase
 {
   private readonly VaultKeepsService _vaultKeepsService;
   private readonly Auth0Provider _auth0Provider;
+  private readonly VaultsService _vaultService;
 
-  public VaultKeepsController(VaultKeepsService vaultKeepsService, Auth0Provider auth0Provider)
+  public VaultKeepsController(VaultKeepsService vaultKeepsService, Auth0Provider auth0Provider, VaultsService vaultsService)
   {
     _vaultKeepsService = vaultKeepsService;
     _auth0Provider = auth0Provider;
+    _vaultService = vaultsService;
   }
 
   [Authorize]
@@ -20,6 +22,11 @@ public class VaultKeepsController : ControllerBase
     try
     {
       Account userInfo = await _auth0Provider.GetUserInfoAsync<Account>(HttpContext);
+      Vault vault = _vaultService.GetVaultById(vaultKeepData.VaultId);
+      if (vault.CreatorId != userInfo.Id)
+      {
+        throw new Exception("This isn't your vault buddy!");
+      }
       vaultKeepData.CreatorId = userInfo.Id;
       vaultKeepData.AccountId = userInfo.Id;
       VaultKeep vaultKeep = _vaultKeepsService.CreateVaultKeep(vaultKeepData);
@@ -38,6 +45,11 @@ public class VaultKeepsController : ControllerBase
     try
     {
       Account userInfo = await _auth0Provider.GetUserInfoAsync<Account>(HttpContext);
+      VaultKeep vaultKeep = _vaultKeepsService.GetVaultKeepById(vaultKeepId);
+      if (vaultKeep.CreatorId != userInfo.Id)
+      {
+        throw new Exception("This isn't your vault buddy!");
+      }
       _vaultKeepsService.RemoveVaultKeep(vaultKeepId, userInfo.Id);
       return Ok("Vault has been closed");
     }
